chore: ghas code scanning issues (#798)
- [ ] New feature
- [ ] Bug fix
- [ ] High impact

**Description of work:**
<!--- Please give a description of the work --->

Fixed security issues that was found when turning on GHAS. Two re
occuring ghas issues fixed and one found while fixing the issues.

- Resources: BeResourceOwnerRequirement and
BeSiblingResourceOwnerRequirement handle suceeded when they should not
have (e.g empty string)
- Moved logic after auth is completed
- Defined permissions in github actions needed for the workflow


[AB#65204](https://statoil-proview.visualstudio.com/787035c2-8cf2-4d73-a83e-bb0e6d937eec/_workitems/edit/65204)

**Testing:**
- [ ] Can be tested
- [ ] Automatic tests created / updated
- [ ] Local tests are passing

<!--- Please give a description of how this can be tested --->


**Checklist:**
- [ ] Considered automated tests
- [ ] Considered updating specification / documentation
- [ ] Considered work items 
- [ ] Considered security
- [ ] Performed developer testing
- [ ] Checklist finalized / ready for review

<!--- Other comments --->

+++ src/backend/api/Fusion.Resources.Api/Controllers/Departments/DepartmentsController.cs
@@ -65,9 +65,6 @@ namespace Fusion.Resources.Api.Controllers
         [ProducesResponseType(StatusCodes.Status200OK)]
         public async Task<ActionResult<ApiDepartment>> GetDepartmentsV11([FromRoute] OrgUnitIdentifier departmentString)
         {
-            if (!departmentString.Exists)
-                return FusionApiError.NotFound(departmentString.OriginalIdentifier, "Department not found");
-
             #region Authorization
 
             var departmentPath = new DepartmentPath(departmentString.FullDepartment);
@@ -87,6 +84,9 @@ namespace Fusion.Resources.Api.Controllers
                 return authResult.CreateForbiddenResponse();
 
             #endregion Authorization
+
+            if (!departmentString.Exists)
+                return FusionApiError.NotFound(departmentString.OriginalIdentifier, "Department not found");
 
             var department = await DispatchAsync(new GetDepartment(departmentString.SapId).ExpandDelegatedResourceOwners());
 
@@ -99,9 +99,6 @@ namespace Fusion.Resources.Api.Controllers
         [ProducesResponseType(StatusCodes.Status204NoContent)]
         public async Task<IActionResult> GetDepartmentAccessV11([FromRoute] OrgUnitIdentifier departmentString)
         {
-            if (!departmentString.Exists)
-                return FusionApiError.NotFound(departmentString.OriginalIdentifier, "Department not found");
-
             #region Authorization
 
             var departmentPath = new DepartmentPath(departmentString.FullDepartment);
@@ -122,6 +119,8 @@ namespace Fusion.Resources.Api.Controllers
 
             #endregion Authorization
 
+            if (!departmentString.Exists)
+                return FusionApiError.NotFound(departmentString.OriginalIdentifier, "Department not found");
 
             var allowedMethods = new List<string>();
             if (authResult.Success)
@@ -154,9 +153,6 @@ namespace Fusion.Resources.Api.Controllers
         [EmulatedUserSupport]
         public async Task<ActionResult> GetDelegatedResourceOwnersOptions([FromRoute] OrgUnitIdentifier departmentString)
         {
-            if (!departmentString.Exists)
-                return FusionApiError.NotFound(departmentString.OriginalIdentifier, "Department not found");
-
             #region Authorization
 
             var authResult = await Request.RequireAuthorizationAsync(r =>
@@ -181,6 +177,9 @@ namespace Fusion.Resources.Api.Controllers
 
             if (authResult.Success)
             {
+                if (!departmentString.Exists)
+                    return FusionApiError.NotFound(departmentString.OriginalIdentifier, "Department not found");
+
                 if (authResult.LimitedAuth == false)
                 {
                     allowedMethods.Add("DELETE");
@@ -192,15 +191,12 @@ namespace Fusion.Resources.Api.Controllers
             Response.Headers["Allow"] = string.Join(',', allowedMethods);
             return NoContent();
         }
-        
+
         [HttpGet("/departments/{departmentString}/delegated-resource-owners")]
         [MapToApiVersion("1.0-preview")]
         [MapToApiVersion("1.0")]
         public async Task<ActionResult<IEnumerable<ApiDepartmentResponsible>>> GetDelegatedDepartmentResponsiblesForDepartment([FromRoute] OrgUnitIdentifier departmentString, bool shouldIgnoreDateFilter)
         {
-            if (!departmentString.Exists)
-                return FusionApiError.NotFound(departmentString.OriginalIdentifier, "Department not found");
-
             #region Authorization
 
             var authResult = await Request.RequireAuthorizationAsync(r =>
@@ -221,6 +217,9 @@ namespace Fusion.Resources.Api.Controllers
 
             #endregion Authorization
 
+            if (!departmentString.Exists)
+                return FusionApiError.NotFound(departmentString.OriginalIdentifier, "Department not found");
+
             if (authResult.Unauthorized)
                 return authResult.CreateForbiddenResponse();
 
@@ -239,9 +238,6 @@ namespace Fusion.Resources.Api.Controllers
         [MapToApiVersion("1.0")]
         public async Task<ActionResult<ApiDepartmentResponsible>> AddDelegatedResourceOwner([FromRoute] OrgUnitIdentifier departmentString, [FromBody] AddDelegatedResourceOwnerRequest request)
         {
-            if (!departmentString.Exists)
-                return FusionApiError.NotFound(departmentString.OriginalIdentifier, "Department not found");
-
             #region Authorization
 
             var authResult = await Request.RequireAuthorizationAsync(r =>
@@ -258,6 +254,9 @@ namespace Fusion.Resources.Api.Controllers
                 return authResult.CreateForbiddenResponse();
 
             #endregion Authorization
+
+            if (!departmentString.Exists)
+                return FusionApiError.NotFound(departmentString.OriginalIdentifier, "Department not found");
 
             try
             {
@@ -291,9 +290,6 @@ namespace Fusion.Resources.Api.Controllers
         [MapToApiVersion("1.0")]
         public async Task<IActionResult> DeleteDelegatedResourceOwner([FromRoute] OrgUnitIdentifier departmentString, Guid azureUniqueId)
         {
-            if (!departmentString.Exists)
-                return FusionApiError.NotFound(departmentString.OriginalIdentifier, "Department not found");
-
             #region Authorization
 
             var authResult = await Request.RequireAuthorizationAsync(r =>
@@ -310,8 +306,11 @@ namespace Fusion.Resources.Api.Controllers
 
             #endregion Authorization
 
+            if (!departmentString.Exists)
+                return FusionApiError.NotFound(departmentString.OriginalIdentifier, "Department not found");
+
             var deleted = await DispatchAsync(new DeleteDelegatedResourceOwner(departmentString.FullDepartment, azureUniqueId));
-            
+
             return NoContent();
         }
